move merkle tree backend trait to merkle module
