+++ crypto/src/hash/mod.rs
@@ -1,4 +1,3 @@
 pub mod hash_to_field;
 pub mod poseidon;
 pub mod sha3;
-pub mod traits;
