+++ crypto/src/hash/sha3/mod.rs
@@ -1,12 +1,5 @@
-use std::marker::PhantomData;
-
-use lambdaworks_math::{
-    field::{element::FieldElement, traits::IsField},
-    traits::ByteConversion,
-};
 use sha3::{Digest, Sha3_256};
 
-use super::traits::IsMerkleTreeBackend;
 pub struct Sha3Hasher;
 
 /// Sha3 Hasher used over fields
@@ -67,43 +60,5 @@ impl Sha3Hasher {
 
     fn strxor(a: &[u8], b: &[u8]) -> Vec<u8> {
         a.iter().zip(b).map(|(a, b)| a ^ b).collect()
-    }
-}
-
-pub struct FieldElementSha3Hasher<F> {
-    phantom: PhantomData<F>,
-}
-
-impl<F> Default for FieldElementSha3Hasher<F> {
-    fn default() -> Self {
-        Self {
-            phantom: PhantomData,
-        }
-    }
-}
-
-impl<F> IsMerkleTreeBackend for FieldElementSha3Hasher<F>
-where
-    F: IsField,
-    FieldElement<F>: ByteConversion,
-{
-    type Node = [u8; 32];
-    type Data = FieldElement<F>;
-
-    fn hash_data(&self, input: &FieldElement<F>) -> [u8; 32] {
-        let mut hasher = Sha3_256::new();
-        hasher.update(input.to_bytes_be());
-        let mut result_hash = [0_u8; 32];
-        result_hash.copy_from_slice(&hasher.finalize());
-        result_hash
-    }
-
-    fn hash_new_parent(&self, left: &[u8; 32], right: &[u8; 32]) -> [u8; 32] {
-        let mut hasher = Sha3_256::new();
-        hasher.update(left);
-        hasher.update(right);
-        let mut result_hash = [0_u8; 32];
-        result_hash.copy_from_slice(&hasher.finalize());
-        result_hash
     }
 }
