+++ crypto/src/merkle_tree/merkle.rs
@@ -1,6 +1,12 @@
-use crate::hash::traits::IsMerkleTreeBackend;
+use std::marker::PhantomData;
 
-use super::{proof::Proof, utils::*};
+use lambdaworks_math::{
+    field::{element::FieldElement, traits::IsField},
+    traits::ByteConversion,
+};
+use sha3::{Digest, Sha3_256};
+
+use super::{proof::Proof, traits::IsMerkleTreeBackend, utils::*};
 
 #[derive(Clone)]
 pub struct MerkleTree<T> {
@@ -57,6 +63,44 @@ where
         }
 
         merkle_path
+    }
+}
+
+pub struct FieldElementBackend<F> {
+    phantom: PhantomData<F>,
+}
+
+impl<F> Default for FieldElementBackend<F> {
+    fn default() -> Self {
+        Self {
+            phantom: PhantomData,
+        }
+    }
+}
+
+impl<F> IsMerkleTreeBackend for FieldElementBackend<F>
+where
+    F: IsField,
+    FieldElement<F>: ByteConversion,
+{
+    type Node = [u8; 32];
+    type Data = FieldElement<F>;
+
+    fn hash_data(&self, input: &FieldElement<F>) -> [u8; 32] {
+        let mut hasher = Sha3_256::new();
+        hasher.update(input.to_bytes_be());
+        let mut result_hash = [0_u8; 32];
+        result_hash.copy_from_slice(&hasher.finalize());
+        result_hash
+    }
+
+    fn hash_new_parent(&self, left: &[u8; 32], right: &[u8; 32]) -> [u8; 32] {
+        let mut hasher = Sha3_256::new();
+        hasher.update(left);
+        hasher.update(right);
+        let mut result_hash = [0_u8; 32];
+        result_hash.copy_from_slice(&hasher.finalize());
+        result_hash
     }
 }
