+++ crypto/src/merkle_tree/mod.rs
@@ -2,4 +2,5 @@ pub mod merkle;
 pub mod proof;
 #[cfg(test)]
 pub mod test_merkle;
+pub mod traits;
 mod utils;
