+++ crypto/src/merkle_tree/traits.rs
