+++ crypto/src/merkle_tree/utils.rs
@@ -1,4 +1,4 @@
-use crate::hash::traits::IsMerkleTreeBackend;
+use super::traits::IsMerkleTreeBackend;
 
 pub fn sibling_index(node_index: usize) -> usize {
     if node_index % 2 == 0 {
@@ -67,7 +67,7 @@ pub fn right_child_index(parent_index: usize) -> usize {
 mod tests {
     use lambdaworks_math::field::{element::FieldElement, fields::u64_prime_field::U64PrimeField};
 
-    use crate::{hash::traits::IsMerkleTreeBackend, merkle_tree::test_merkle::TestHasher};
+    use crate::merkle_tree::{test_merkle::TestHasher, traits::IsMerkleTreeBackend};
 
     use super::{build, complete_until_power_of_two};
